backend/requirements.txt
# Core Framework
fastapi==0.121.3
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
alembic==1.17.2
psycopg[binary]>=3.2.0
asyncpg==0.30.0
pgvector==0.3.6

# Data Validation
pydantic>=2.10.5
pydantic-settings>=2.7.1

# AI/ML
sentence-transformers==5.1.2
torch==2.6.0
langchain==0.3.13
langchain-community==0.3.13
tiktoken==0.8.0
rank-bm25==0.2.2
numpy>=1.24.0
scipy>=1.10.0  # For embedding quality analysis

# File Processing
PyPDF2==3.0.1
pymupdf==1.25.2
python-magic==0.4.27
pillow==11.0.0

# Image Processing
open-clip-torch==3.2.0
easyocr==1.7.2

# Audio Processing
faster-whisper==1.1.0
soundfile==0.12.1
librosa==0.10.2.post1

# Object Storage
minio==7.2.14

# Utilities
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx>=0.27.0,<0.28.0
trafilatura==1.12.2
beautifulsoup4==4.12.3
sse-starlette==2.2.1
ollama==0.4.4
slowapi==0.1.9

# Logging & Monitoring
structlog==24.4.0
python-json-logger==3.2.1

# CORS
fastapi-cors==0.0.6